# api/docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Engagement
  version: "2.0"
  description: |
    API partenaires pour la diffusion et le suivi de missions bénévoles et civiques.

    ## Authentification
    Tous les endpoints nécessitent une clé API transmise dans le header "x-api-key".

    ## Versions
    - **v0** : endpoints de lecture (recherche de missions, organisations, partenaires diffuseurs)
    - **v2** : endpoints d'écriture CRUD (création/modification/suppression de missions, suivi des activités)
  license:
    name: MIT
servers:
  - url: https://api.api-engagement.beta.gouv.fr
    description: Production
  - url: https://api.bac-a-sable.api-engagement.beta.gouv.fr
    description: Bac à sable
  - url: http://localhost:3002
    description: Développement local
security:
  - ApiKeyAuth: []
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Clé API fournie par l'équipe API Engagement
  schemas:
...
    Address:
      type: object
      description: Adresse géographique d'une mission
      properties:
        street:
          type: string
          nullable: true
          description: Numéro et nom de rue
          example: "12 rue de la Paix"
        postalCode:
          type: string
          nullable: true
          description: Code postal
          example: "75001"
        city:
          type: string
          nullable: true
          description: Ville
          example: "Paris"
        departmentCode:
          type: string
          nullable: true
          description: Code du département (ex. 75, 69, 2A)
          example: "75"
        departmentName:
          type: string
          nullable: true
          description: Nom du département
          example: "Paris"
        region:
          type: string
          nullable: true
          description: Nom de la région
          example: "Île-de-France"
        country:
          type: string
          nullable: true
          description: Pays (code ISO 3166-1 alpha-2 ou nom)
          example: "France"

    # ── Mission (réponse v2) ───────────────────────────────────────────────────
...
    MissionV2:
      description: Mission bénévole ou civique (format v2, étend Mission)
      allOf:
        - $ref: "#/components/schemas/Mission"
        - type: object
          properties:
            openToMinors:
              type: boolean
              nullable: true
              description: Mission ouverte aux mineurs.
            reducedMobilityAccessible:
              type: boolean
              nullable: true
              description: Mission accessible aux personnes à mobilité réduite.
            closeToTransport:
              type: boolean
              nullable: true
              description: Mission proche des transports en commun.

    # ── Mission (réponse v0 — format legacy) ──────────────────────────────────
...
  responses:
    BadRequest:
      description: Paramètres invalides
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            invalid_body:
              value:
                ok: false
                code: "INVALID_BODY"
                message: "clientId is required"
            invalid_params:
              value:
                ok: false
                code: "INVALID_PARAMS"
                message: "Invalid path parameter"
    Unauthorized:
      description: Clé API manquante ou invalide
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            ok: false
            code: "UNAUTHORIZED"
            message: "Unauthorized"
    NotFound:
      description: Ressource introuvable
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            ok: false
            code: "NOT_FOUND"
            message: "Mission not found"
...
    Conflict:
      description: Conflit — la ressource existe déjà
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            ok: false
            code: "RESSOURCE_ALREADY_EXIST"
            message: "A mission with this clientId already exists for this publisher"
...
paths:
...
  /v2/mission:
    post:
      summary: Créer une mission
      description: |
        Crée une nouvelle mission pour votre organisation.

        La mission est immédiatement soumise aux règles de modération automatique :
        si elle est incomplète (description trop courte, pas d'adresse, etc.), le champ
        "statusCode" sera positionné à "REFUSED" avec un commentaire explicatif.

        > **Rate limiting** : endpoint soumis à limitation de débit.
      operationId: createMission
      tags: [Missions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [clientId, title]
              properties:
                clientId:
                  type: string
                  description: Identifiant unique de la mission dans votre système d'information (immuable)
                  example: "mission-2024-001"
                title:
                  type: string
                  description: Titre de la mission
                  example: "Animateur bénévole en maison de retraite"
                description:
                  type: string
                  description: |
                    Description complète de la mission. Texte brut ou HTML acceptés, stockés tels quels.

                    **Recommandations** :
                    - Utiliser le format HTML pour une meilleure lisibilité lors de la diffusion (sauts de ligne, gras, listes)
                    - Structurer en plusieurs sections pour agréger des informations provenant de champs différents dans votre système
                    - La description est limitée à 20 000 caractères
                  example: "<p><strong>EN QUELQUES MOTS</strong></p><p>Favoriser la mixité au sein du club</p><p><strong>VOTRE MISSION DE SERVICE CIVIQUE</strong></p><p><u>Objectifs</u></p><ul><li>Développer la féminisation et favoriser l'accès à la pratique du football des jeunes filles</li></ul><p><u>Actions</u></p><ul><li>Accompagner les éducateurs dans le projet éducatif féminin</li><li>Mettre en place des actions en faveur de la mixité</li><li>Sensibiliser les jeunes filles éloignées de la pratique sportive</li></ul>"
                applicationUrl:
                  type: string
                  format: uri
                  description: URL de la page où l'utilisateur sera redirigé pour candidater (URL de la page présentant la mission)
                  example: "https://www.example.org/missions/animateur-maison-retraite"
                image:
                  type: string
                  format: uri
                  description: URL de l'image de la mission. Si absent, l'API Engagement utilise une image de sa bibliothèque selon le domaine.
                metadata:
                  type: string
                  description: Métadonnées libres associées à la mission.
                domain:
                  type: string
                  description: |
                    Domaine d'action de la mission. Valeurs possibles :
                    `animaux`, `autre`, `batiment-industrie-logistique`, `benevolat-competences`, `communication`, `culture-loisirs`, `education`, `emploi`, `environnement`, `gestion-finance-droit`, `humanitaire`, `memoire-et-citoyennete`, `numerique`, `prevention-protection`, `recherche`, `sante`, `service-public-defense-securite`, `sport`, `solidarite-insertion`, `vivre-ensemble`.

                    > Une demande d'ajout de valeur peut être faite auprès de l'équipe API Engagement.
                  example: "sante"
                activities:
                  type: array
                  items:
                    type: string
                  description: |
                    Activité(s) principale(s) de la mission. Valeurs possibles :
                    `accueil-de-public`, `aide-psychologique`, `activites-manuelles`, `alphabetisation`, `animation`, `Animation, Valorisation`, `art`, `autre`, `bricolage`, `collecte`, `communication`, `comptabilite-finance`, `conseil`, `distribution`, `documentation-traduction`, `ecoute-permanence`, `encadrement-d-equipes`, `enseignement-formation`, `gestion-de-projets`, `gestion-recherche-des-partenariats`, `informatique`, `jardinage`, `juridique`, `logistique`, `lutte-contre-isolement`, `Médiation, Information`, `mentorat-parrainage`, `mission-internationale`, `Préservation, Patrimoine`, `Prévention, Sensibilisation`, `ramassage-dechets`, `recrutement`, `responsabilites-associatives`, `sante-soins`, `Secours, Aide`, `secourisme`, `sensibilisation`, `soins-animaux`, `Soutien, Accompagnement`, `soutien-scolaire`, `sport`, `taches-administratives`, `Transmission, Pédagogie`, `visites`, `operations-militaires`, `sciences-humaines-social`, `restauration-hotellerie`, `maintenance-technique`, `gestion-ressources-humaines`, `gestion-entretien-patrimoine`.

                    > Une demande d'ajout de valeur peut être faite auprès de l'équipe API Engagement.
                  example: ["animation", "lutte-contre-isolement"]
                tags:
                  type: array
                  items:
                    type: string
                  description: Mots-clés personnalisés pour taguer la mission. Chaque élément du tableau est un tag distinct.
                  example: ["Ecologie", "Sport", "Solidarité"]
                audience:
                  type: array
                  items:
                    type: string
                  description: |
                    Les bénéficiaires de la mission — à qui la mission s'adresse, le public auquel sera confronté la personne qui s'engage.
                  example: ["Tous publics"]
                requirements:
                  type: array
                  items:
                    type: string
                  description: Pré-requis relatifs à la mission.
                  example: ["Niveau d'étude : BAC+2", "Connaissance de langue anglaise"]
                softSkills:
                  type: array
                  items:
                    type: string
                  description: |
                    Les compétences générales requises pour la mission. Ce champ ne repose pas sur un référentiel de compétence structuré.
                  example: ["Écoute active", "Communication orale"]
                romeSkills:
                  type: array
                  items:
                    type: string
                  description: |
                    Code OGR Macro du ROME 4.0. Ce champ repose sur le référentiel "Arborescence simplifiée des compétences" (section "Arborescence du ROME").
                  example: ["300412", "300361"]
                remote:
                  type: string
                  enum: [no, possible, full]
                  description: "Disponibilité en distanciel. Valeurs possibles : `no` (présentiel uniquement), `possible` (distanciel possible), `full` (100% distanciel)"
                  example: "no"
                schedule:
                  type: string
                  description: Rythme de la mission (texte libre)
                  example: "2 heures par semaine"
                startAt:
                  type: string
                  format: date-time
                  description: "Date de début de la mission souhaitée. Format ISO 8601. Défaut : now()"
                  example: "2024-09-01T00:00:00.000Z"
                endAt:
                  type: string
                  format: date-time
                  description: "Date de fin de la mission souhaitée. Format ISO 8601. Défaut : now()"
                  example: "2025-06-30T00:00:00.000Z"
                postedAt:
                  type: string
                  format: date-time
                  description: "Date de première publication de la mission. Format ISO 8601."
                  example: "2024-06-15T09:00:00.000Z"
                places:
                  type: integer
                  minimum: 1
                  default: 1
                  description: |
                    Nombre de places disponibles pour la mission. Ce nombre doit être mis à jour dès qu'il évolue dans votre système afin que l'API Engagement dispose toujours de la donnée la plus récente.
                compensationAmount:
                  type: number
                  description: Montant minimum de l'indemnisation. Utilisé seul ou comme borne basse d'une fourchette avec `compensationAmountMax`.
                  example: 0
                compensationAmountMax:
                  type: number
                  description: Montant maximum de l'indemnisation. Si renseigné, doit être supérieur ou égal à `compensationAmount` (ex. 0–13 €/h).
                  example: 13
                compensationUnit:
                  type: string
                  enum: [hour, day, month, year]
                  description: "Période de l'indemnisation pour le montant fourni. Valeurs possibles : `year`, `month`, `day`, `hour`"
                compensationType:
                  type: string
                  enum: [gross, net]
                  description: "Type d'indemnisation. Valeurs possibles : `gross` (brut), `net` (net)"
                openToMinors:
                  type: boolean
                  description: Mission ouverte aux mineurs.
                reducedMobilityAccessible:
                  type: boolean
                  description: Mission accessible aux personnes à mobilité réduite.
                closeToTransport:
                  type: boolean
                  description: Mission proche des transports en commun.
                addresses:
                  type: array
                  items:
                    $ref: "#/components/schemas/Address"
                  description: |
                    Ensemble des localisations où se déroule la mission. Si la mission se tient dans plusieurs lieux, renseignez autant d'objets que de localisations. Pour une mission en lieu unique, le tableau ne contient qu'un seul objet.
                  example:
                    - street: "12 rue de la Paix"
                      postalCode: "75001"
                      city: "Paris"
                      departmentCode: "75"
                      departmentName: "Paris"
                      region: "Île-de-France"
                      country: "France"
                type:
                  type: string
                  enum: [benevolat, volontariat_service_civique, volontariat_sapeurs_pompiers, volontariat_reserve_operationnelle]
                  description: Type de mission d'engagement
                organizationClientId:
                  type: string
                  description: Identifiant de l'organisation dans votre système. L'organisation désigne l'entité qui accueille le bénévole ou le volontaire.
                organizationName:
                  type: string
                  description: "Nom de l'organisation. Requis si un autre champ `organization*` est renseigné."
                organizationDescription:
                  type: string
                  description: Description de l'organisation.
                organizationUrl:
                  type: string
                  format: uri
                  description: Lien de l'organisation (ex. page de présentation sur votre site).
                organizationType:
                  type: string
                  description: Type de l'organisation.
                  example: "Association déclarée"
                organizationLogo:
                  type: string
                  format: uri
                  description: URL de l'image de l'organisation.
                organizationRNA:
                  type: string
                  description: Numéro RNA de l'organisation. Uniquement valable pour les organisations de type association.
                  example: "W353002476"
                organizationSiren:
                  type: string
                  description: SIREN de l'organisation.
                  example: "339863417"
                organizationSiret:
                  type: string
                  description: SIRET de l'organisation.
                  example: "33986341700418"
                organizationFullAddress:
                  type: string
                  description: Adresse de l'organisation. Pas de format strict attendu (contrairement à l'adresse de la mission).
                organizationPostCode:
                  type: string
                  description: Code postal de l'organisation.
                  example: "94170"
                organizationCity:
                  type: string
                  description: Ville de l'organisation.
                  example: "Le Perreux-sur-Marne"
                organizationDepartmentCode:
                  type: string
                  description: Code du département de l'organisation.
                  example: "94"
                organizationDepartmentName:
                  type: string
                  description: Nom du département de l'organisation.
                  example: "Val de Marne"
                organizationStatusJuridique:
                  type: string
                  description: Statut juridique de l'organisation.
                organizationBeneficiaries:
                  type: array
                  items:
                    type: string
                  description: Bénéficiaires de l'organisation.
                organizationActions:
                  type: array
                  items:
                    type: string
                  description: Principales actions menées par l'organisation.
                  example: ["Animation", "Valorisation", "Préservation"]
                organizationReseaux:
                  type: array
                  items:
                    type: string
                  description: Réseau de l'organisation. Ce champ sert à renseigner les organisations ayant des antennes locales et dépendant d'un réseau national.
      responses:
        "201":
          description: Mission créée
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/MissionV2"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /v0/mission/search:
    get:
      summary: Rechercher des missions avec facettes
      description: |
        Identique à "GET /v0/mission" mais retourne également des facettes
        (agrégations) sur les domaines, activités et départements pour faciliter
        la construction de filtres dynamiques.
      operationId: searchMissionsWithFacets
      tags: [Missions]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 10000
            default: 25
        - name: skip
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: keywords
          in: query
          schema:
            type: string
        - name: domain
          in: query
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: activity
          in: query
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: type
          in: query
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: remote
          in: query
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: lat
          in: query
          schema:
            type: number
            format: double
        - name: lon
          in: query
          schema:
            type: number
            format: double
        - name: distance
          in: query
          schema:
            type: string
        - name: departmentName
          in: query
          schema:
            oneOf:
              - type: string
              - type: array
                items:
                  type: string
        - name: openToMinors
          in: query
          schema:
            type: string
            enum: ["true", "false", "yes", "no", "1", "0"]
        - name: reducedMobilityAccessible
          in: query
          schema:
            type: string
            enum: ["true", "false", "yes", "no", "1", "0"]
      responses:
        "200":
          description: Liste de missions avec facettes
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  total:
                    type: integer
                    example: 1234
                  hits:
                    type: array
                    items:
                      $ref: "#/components/schemas/MissionLegacy"
                  facets:
                    type: object
                    description: Agrégations pour les filtres dynamiques
                    properties:
                      domains:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            doc_count:
                              type: integer
                      activities:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            doc_count:
                              type: integer
                      departmentName:
                        type: array
                        items:
                          type: object
                          properties:
                            key:
                              type: string
                            doc_count:
                              type: integer
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v0/mission/{id}:
    get:
      summary: Récupérer une mission par ID
      operationId: getMission
      tags: [Missions]
      parameters:
        - name: id
          in: path
          required: true
          description: Identifiant interne de la mission
          schema:
            type: string
          example: "64a1b2c3d4e5f6789abc0001"
      responses:
        "200":
          description: Mission trouvée
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    example: true
                  data:
                    $ref: "#/components/schemas/MissionLegacy"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

  # ────────────────────────────────────────────────────────────────────────────
  # v0 — Mes missions
  # ────────────────────────────────────────────────────────────────────────────
...